rest data api

+++ flux/src/main/java/web/flux/repository/UserRepository.java
@@ -3,10 +3,12 @@ package web.flux.repository;
 import web.flux.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.math.BigInteger;
+
 /**
  * @author z
  */
-public interface UserRepository extends JpaRepository<User, Long>{
+public interface UserRepository extends JpaRepository<User, BigInteger>{
     /**
      * spring security
      * @param username username
